split unknown types test
