+++ tests/Type/Nette/data/json_decode.php
@@ -20,11 +20,3 @@ assertType('stdClass', $value);
 
 $value = Json::decode('[1, 2, 3]');
 assertType('array{1, 2, 3}', $value);
-
-
-function unknownType($mixed) {
-	$value = Json::decode($mixed);
-	assertType('bool|float|int|stdClass|string', $value);
-}
-
-
