+++ tests/Type/Nette/data/json_decode_force_array_unknown_type.php
@@ -1,0 +1,9 @@
+<?php
+
+use Nette\Utils\Json;
+use function PHPStan\Testing\assertType;
+
+function unknownType($mixed) {
+	$value = Json::decode($mixed, Json::FORCE_ARRAY);
+	assertType('array|bool|float|int|string', $value);
+}
