+++ tests/Type/Nette/data/json_decode_unknown_type.php
@@ -1,0 +1,9 @@
+<?php
+
+use Nette\Utils\Json;
+use function PHPStan\Testing\assertType;
+
+function unknownType($mixed) {
+	$value = Json::decode($mixed);
+	assertType('bool|float|int|stdClass|string', $value);
+}
